Fix rating
